# Copy player breast settings to armor stands