--- a/src/main/java/com/wildfire/main/WildfireHelper.java
+++ b/src/main/java/com/wildfire/main/WildfireHelper.java
@@ -20,18 +20,25 @@ package com.wildfire.main;
 
 import com.wildfire.api.IGenderArmor;
 import com.wildfire.api.WildfireAPI;
+import com.wildfire.main.entitydata.Breasts;
+import com.wildfire.main.entitydata.EntityConfig;
+import com.wildfire.main.entitydata.PlayerConfig;
 import com.wildfire.render.armor.SimpleGenderArmor;
 import com.wildfire.render.armor.EmptyGenderArmor;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.font.TextRenderer;
 import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.render.entity.PlayerModelPart;
 import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.*;
+import net.minecraft.nbt.NbtCompound;
 import net.minecraft.text.Text;
 import net.minecraft.util.Util;
 import net.minecraft.util.math.MathHelper;
 
+import javax.annotation.Nonnull;
 import java.util.Objects;
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -87,6 +94,7 @@ public class WildfireHelper {
         ctx.drawText(textRenderer, text, centeredX, y, color, false);
     }
 
+    @Environment(EnvType.CLIENT)
     public static void drawScrollableText(DrawContext context, TextRenderer textRenderer, Text text, int left, int top, int right, int bottom, int color) {
         int i = textRenderer.getWidth(text);
         int var10000 = top + bottom;
@@ -105,6 +113,33 @@ public class WildfireHelper {
         } else {
             drawCenteredText(context, textRenderer, text, (left + right) / 2, j, color);
         }
+    }
 
+    /**
+     * <p>Write a player's gender config to NBT on the given item stack.</p>
+     *
+     * <p>This only copies enough data to render breasts similarly to how they'd appear on the given player, which includes:</p>
+     * <ul>
+     *     <li>{@link EntityConfig#getBustSize() Breast size}</li>
+     *     <li>{@link Breasts#getCleavage() Cleavage}</li>
+     *     <li>{@link Breasts#isUniboob() Uniboob}</li>
+     *     <li>{@link Breasts#getXOffset() X}, {@link Breasts#getYOffset() Y}, and {@link Breasts#getZOffset() Z} offsets</li>
+     *     <li>Whether the {@link PlayerEntity#isPartVisible player's jacket layer is visible}</li>
+     * </ul>
+     *
+     * @see EntityConfig#readFromStack
+     */
+    public static void writeToNbt(@Nonnull PlayerEntity player, @Nonnull PlayerConfig config, @Nonnull ItemStack armor) {
+        NbtCompound nbt = new NbtCompound();
+        nbt.putFloat("BreastSize", config.getGender().canHaveBreasts() && config.showBreastsInArmor() ? config.getBustSize() : 0f);
+        nbt.putFloat("Cleavage", config.getBreasts().getCleavage());
+        nbt.putBoolean("Uniboob", config.getBreasts().isUniboob());
+        nbt.putFloat("XOffset", config.getBreasts().getXOffset());
+        nbt.putFloat("YOffset", config.getBreasts().getYOffset());
+        nbt.putFloat("ZOffset", config.getBreasts().getZOffset());
+        // note that we also copy this to properly copy the exact size, as the player model will push the breast armor
+        // layer out a bit if they have a visible jacket layer
+        nbt.putBoolean("Jacket", player.isPartVisible(PlayerModelPart.JACKET));
+        armor.setSubNbt("WildfireGender", nbt);
     }
 }