--- a/src/main/java/com/wildfire/render/BreastSide.java
+++ b/src/main/java/com/wildfire/render/BreastSide.java
@@ -1,0 +1,5 @@
+package com.wildfire.render;
+
+public enum BreastSide {
+	LEFT, RIGHT
+}