… src/main/java/com/wildfire/main/entitydata/Breasts.java
… src/main/java/com/wildfire/main/entitydata/EntityConfig.java
… src/main/java/com/wildfire/main/entitydata/PlayerConfig.java
… src/main/java/com/wildfire/main/networking/WildfireSync.java
… src/main/java/com/wildfire/mixins/ArmorStandEntityMixin.java
… src/main/java/com/wildfire/mixins/ArmorStandEntityRendererMixin.java
… src/main/java/com/wildfire/mixins/BreastPhysicsTickMixin.java
… src/main/java/com/wildfire/mixins/LivingEntityMixin.java
… src/main/java/com/wildfire/physics/BreastPhysics.java
… src/main/java/com/wildfire/render/GenderArmorLayer.java
… src/main/java/com/wildfire/render/GenderLayer.java
… src/main/java/com/wildfire/render/armor/SimpleGenderArmor.java
… src/main/resources/wildfire_gender.mixins.json