(I'd like to preface this by deeply apologizing for just how borderline unreviewable these rendering changes are)

This implements [this suggestion](https://canary.discord.com/channels/820872727539875841/1177844764926820373), along with the foundations to allow for [rendering on most other other mobs](https://canary.discord.com/channels/820872727539875841/1167685071810089100).

The main changes here include splitting rendering into two different renderers; one for the skin layer, and another for the armor layer, which mirrors how vanilla handles rendering armor.

Along with the above, this also splits the `GenderPlayer` class into two classes; a new `EntityConfig`, which serves as a stripped-down version of the full - now renamed - `PlayerConfig`, simply providing the bare minimum to allow for rendering and physics simulation.

The way that this attaches the player's settings to the armor stand is by making use of an NBT tag on the chestplate; this implementation does have the (very slight) benefit of *technically* not requiring the mod server-side for this feature, even if it's not very relevant (unless you have access to `/give`, that is).

Note that this may have performance issues I haven't actually been able to benchmark, due to just how often armor stands may be used as markers and/or text displays on servers/custom maps, although I have tried to keep any such impacts to a minimum by failing fast wherever reasonable.