@WildfireRomeo 2023-11-27
Okay, this is really cool NGL... I'm definitely interested in merging this. Although, as you said, there may be performance implications with it. Haven't tested it extensively.

Has it been tested in multiplayer? How does it react when a client has the mod and the server doesn't? Does it just not show the settings or will it only show for that specific client?