@celestialfault 2023-11-27
> How does it react when a client has the mod and the server doesn't?

The settings simply won't show; this requires that the mod is also installed server-side, as the client can't attach the necessary NBT data before the armor is put on the armor stand.