@WildfireRomeo 2023-11-28
### Tested 147 Armor Stands
It's not a perfect benchmark as there's external factors of framerate affecting in this world, but it gives an idea of how performance would change. I might try to do another benchmark on a flat world, but I don't see it necessary.

It's honestly not as much of a performance drop as I thought it would be... Compared to the game just rendering armor stands with a chestplate on it with the breasts rendering (14 fps drop). This seems totally acceptable to myself, as I don't see many people having over 100 armor stands in any given area with all of them rendering fully (maybe invisible ones, but that would increase the performance).

Without Armor (176 fps)
![2023-11-27_22 43 45](https://github.com/WildfireRomeo/WildfireFemaleGenderMod/assets/80865771/97b80a30-c052-4b73-9ee3-cb2200417a7c)

With Armor - without breasts rendering (107 fps)
![image](https://github.com/WildfireRomeo/WildfireFemaleGenderMod/assets/80865771/b7093cf6-27f1-4bbd-ab3f-c0eab63554b2)


With Armor - with breasts rendering (93 fps)
![2023-11-27_22 44 59](https://github.com/WildfireRomeo/WildfireFemaleGenderMod/assets/80865771/92cb0074-8954-463c-a5fb-6b15fb596e5a)